Basic Firebase integration and testing, restructured riectory structure
