+++ app/_layout.tsx
@@ -1,17 +1,23 @@
 import { Stack } from 'expo-router';
 import React from 'react';
-import { Colors } from './constants';
 
 export default function RootLayout() {
   return (
-    <Stack
-      screenOptions={{
-        headerShown: false,
-        contentStyle: { backgroundColor: Colors.background.secondary },
-      }}
-    >
-      <Stack.Screen name="index" />
-      <Stack.Screen name="enter-code" />
+    <Stack>
+      <Stack.Screen 
+        name="index" 
+        options={{ 
+          title: 'LaterGram',
+          headerShown: false 
+        }} 
+      />
+      <Stack.Screen 
+        name="enter-code" 
+        options={{ 
+          title: 'Enter Event Code',
+          headerBackTitle: 'Back'
+        }} 
+      />
     </Stack>
   );
 }
