+++ src/components/constants/Colors.ts
