+++ src/components/constants/Spacing.ts
@@ -1,4 +1,3 @@
-// app/constants/Spacing.ts
 export const Spacing = {
   xs: 4,
   sm: 8,
