+++ src/components/ui/Card.tsx
@@ -1,6 +1,6 @@
+import { BorderRadius, Colors, Spacing } from '@/components/constants';
 import React from 'react';
 import { StyleSheet, View, ViewStyle } from 'react-native';
-import { BorderRadius, Colors, Spacing } from '../../../constants';
 
 interface CardProps {
   children: React.ReactNode;
@@ -35,13 +35,7 @@ const styles = StyleSheet.create({
     borderColor: Colors.border.light,
   },
   elevated: {
-    shadowColor: Colors.neutral[900],
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 8,
-    elevation: 4,
+    boxShadow: '0px 2px 8px rgba(0, 0, 0, 0.1)',
   },
 });
+
